Maya separated from main.py The menu-item is being added directly within the userSetup.py where it belongs (though using quite the hacky implementation, do have a look to see if you can improve).
